+++ modules/src/ics03_connection/handler/verify.rs
@@ -197,12 +197,9 @@ pub fn check_client_consensus_height(
         return Err(Kind::InvalidConsensusHeight(claimed_height, ctx.host_current_height()).into());
     }
 
-    let oldest_available_height =
-        ctx.host_current_height().revision_height - ctx.host_chain_history_size() as u64;
-
-    if claimed_height.revision_height < oldest_available_height {
+    if claimed_height < ctx.host_oldest_height() {
         // Fail if the consensus height is too old (has been pruned).
-        return Err(Kind::StaleConsensusHeight(claimed_height, ctx.host_current_height()).into());
+        return Err(Kind::StaleConsensusHeight(claimed_height, ctx.host_oldest_height()).into());
     }
 
     // Height check is within normal bounds, check passes.